refactor: rename middleware follow go convention
