+++ internal/middleware/auth.go
@@ -1,4 +1,4 @@
-package middlewares
+package middleware
 
 import (
 	"errors"
